deploy Sprint: Nov 29 - Dec 12, 2022

+++ backend/src/main/java/ch/puzzle/okr/dto/ObjectiveDto.java
@@ -12,10 +12,11 @@ public class ObjectiveDto {
     private Integer quarterNumber;
     private Integer quarterYear;
     private String description;
-    private Double progress;
+    private Long progress;
 
-    public ObjectiveDto(Long id, String title, Long ownerId, String ownerFirstname, String ownerLastname, Long teamId, String teamName,
-                        Long quarterId, Integer quarterNumber, Integer quarterYear, String description, Double progress) {
+    public ObjectiveDto(Long id, String title, Long ownerId, String ownerFirstname, String ownerLastname, Long teamId,
+            String teamName, Long quarterId, Integer quarterNumber, Integer quarterYear, String description,
+            Long progress) {
         this.id = id;
         this.title = title;
         this.ownerId = ownerId;
@@ -29,7 +30,6 @@ public class ObjectiveDto {
         this.description = description;
         this.progress = progress;
     }
-
 
     public Long getId() {
         return id;
@@ -95,7 +95,6 @@ public class ObjectiveDto {
         this.quarterId = quarterId;
     }
 
-
     public Integer getQuarterNumber() {
         return quarterNumber;
     }
@@ -120,11 +119,11 @@ public class ObjectiveDto {
         this.description = description;
     }
 
-    public Double getProgress() {
+    public Long getProgress() {
         return progress;
     }
 
-    public void setProgress(Double progress) {
+    public void setProgress(Long progress) {
         this.progress = progress;
     }
 }
