+++ backend/src/main/java/ch/puzzle/okr/dto/goal/GoalDto.java
@@ -9,7 +9,7 @@ public class GoalDto {
     private GoalKeyResultDto keyresult;
     private Long teamId;
     private String teamName;
-    private Double progress;
+    private Long progress;
     private Integer quarterNumber;
     private Integer quarterYear;
     private ExpectedEvolution expectedEvolution;
@@ -17,9 +17,9 @@ public class GoalDto {
     private Long basicValue;
     private Long targetValue;
 
-    public GoalDto(GoalObjectiveDto objective, GoalKeyResultDto keyresult, Team team, Double progress,
-                   Integer quarterNumber, Integer quarterYear, ExpectedEvolution expectedEvolution, Unit unit,
-                   Long basicValue, Long targetValue) {
+    public GoalDto(GoalObjectiveDto objective, GoalKeyResultDto keyresult, Team team, Long progress,
+            Integer quarterNumber, Integer quarterYear, ExpectedEvolution expectedEvolution, Unit unit, Long basicValue,
+            Long targetValue) {
         this.objective = objective;
         this.keyresult = keyresult;
         this.teamId = team.getId();
@@ -65,11 +65,11 @@ public class GoalDto {
         this.teamName = teamName;
     }
 
-    public Double getProgress() {
+    public Long getProgress() {
         return progress;
     }
 
-    public void setProgress(Double progress) {
+    public void setProgress(Long progress) {
         this.progress = progress;
     }
